do not take a screenshot of column menu components list
